requirements.txt
streamlit>=1.28.0
chromadb>=0.4.15
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-mock>=3.11.0
PyMuPDF>=1.23.0
Pillow>=9.0.0
openai
langchain
langchain-openai
langchain-chroma
langchain-core